requirements.txt
pymodbus
paho.mqtt